Reprogram PowerShell Scripts
Reprogrammed PowerShell Script, as well as added a new one, that allows for a file or folder to be delivered alongside the script. This is opposed to how the original script worked, as it just assumed whatever parent folder the script resided in was what should be added to the PATH. It is my intention that these scripts be run via a context menu shortcut within file explorer, hence the change.